moving struct & handling dwm transparency on backend
