+++ pyglet/libs/win32/constants.py
@@ -5087,3 +5087,8 @@ MF_FILEFLAGS_NONE = 0
 MF_FILEFLAGS_NOBUFFERING = 1
 
 CLSCTX_INPROC_SERVER = 0x1
+
+# From Dwmapi.h
+DWM_BB_ENABLE = 0x00000001
+DWM_BB_BLURREGION = 0x00000002
+DWM_BB_TRANSITIONONMAXIMIZED = 0x00000004
